Refactor single-threaded benchmarks to test all ART flavors
- Parameterize all benchmarks on the ART class, add mutex_db and olc_db as
  benchmark parameters everywhere.
- Cleanup micro_benchmark_mutex.cpp to use micro_benchmark_concurrency.hpp
  facilities
- Split away micro_benchmark_node_utils.hpp from micro_benchmark_utils.hpp and
  micro_benchmark_utils.cpp so that non-node-level benchmarks (such as
  micro_benchmark_key_prefix and the concurrent benchmarks) do not have to
  compile all the node-level benchmark machinery.
- At the same time remove extern template compilation and define node benchmark
  utility function templates in the header due to extern template instantiation
  matrix parameter space expansion.
- Inline unodb::benchmark::get_key into its sole caller.
- Tune quick_benchmark filters
