+++ benchmark/micro_benchmark_node256.cpp
@@ -1,82 +1,166 @@
-// Copyright 2020 Laurynas Biveinis
+// Copyright 2020-2021 Laurynas Biveinis
 
 #include "global.hpp"
 
 #include <benchmark/benchmark.h>
 
 #include "art.hpp"
+#include "micro_benchmark_node_utils.hpp"
 #include "micro_benchmark_utils.hpp"
+#include "mutex_art.hpp"
+#include "olc_art.hpp"
 
 namespace {
 
+template <class Db>
 void grow_node48_to_node256_sequentially(benchmark::State &state) {
-  unodb::benchmark::grow_node_sequentially_benchmark<unodb::db, 48>(state);
+  unodb::benchmark::grow_node_sequentially_benchmark<Db, 48>(state);
 }
 
+template <class Db>
 void grow_node48_to_node256_randomly(benchmark::State &state) {
-  unodb::benchmark::grow_node_randomly_benchmark<unodb::db, 48>(state);
+  unodb::benchmark::grow_node_randomly_benchmark<Db, 48>(state);
 }
 
+template <class Db>
 void node256_sequential_add(benchmark::State &state) {
-  unodb::benchmark::sequential_add_benchmark<unodb::db, 256>(state);
+  unodb::benchmark::sequential_add_benchmark<Db, 256>(state);
 }
 
+template <class Db>
 void node256_random_add(benchmark::State &state) {
-  unodb::benchmark::random_add_benchmark<unodb::db, 256>(state);
+  unodb::benchmark::random_add_benchmark<Db, 256>(state);
 }
 
+template <class Db>
 void minimal_node256_tree_full_scan(benchmark::State &state) {
-  unodb::benchmark::minimal_tree_full_scan<unodb::db, 256>(state);
+  unodb::benchmark::minimal_tree_full_scan<Db, 256>(state);
 }
 
+template <class Db>
 void minimal_node256_tree_random_gets(benchmark::State &state) {
-  unodb::benchmark::minimal_tree_random_gets<unodb::db, 256>(state);
+  unodb::benchmark::minimal_tree_random_gets<Db, 256>(state);
 }
 
+template <class Db>
 void full_node256_tree_full_scan(benchmark::State &state) {
-  unodb::benchmark::full_node_scan_benchmark<unodb::db, 256>(state);
+  unodb::benchmark::full_node_scan_benchmark<Db, 256>(state);
 }
 
+template <class Db>
 void full_node256_tree_random_gets(benchmark::State &state) {
-  unodb::benchmark::full_node_random_get_benchmark<unodb::db, 256>(state);
+  unodb::benchmark::full_node_random_get_benchmark<Db, 256>(state);
 }
 
+template <class Db>
 void full_node256_tree_sequential_delete(benchmark::State &state) {
-  unodb::benchmark::sequential_delete_benchmark<unodb::db, 256>(state);
+  unodb::benchmark::sequential_delete_benchmark<Db, 256>(state);
 }
 
+template <class Db>
 void full_node256_tree_random_delete(benchmark::State &state) {
-  unodb::benchmark::random_delete_benchmark<unodb::db, 256>(state);
+  unodb::benchmark::random_delete_benchmark<Db, 256>(state);
 }
 
 }  // namespace
 
-BENCHMARK(grow_node48_to_node256_sequentially)
+BENCHMARK_TEMPLATE(grow_node48_to_node256_sequentially, unodb::db)
     ->Range(2, 2048)
     ->Unit(benchmark::kMicrosecond);
-BENCHMARK(grow_node48_to_node256_randomly)
+BENCHMARK_TEMPLATE(grow_node48_to_node256_sequentially, unodb::mutex_db)
     ->Range(2, 2048)
     ->Unit(benchmark::kMicrosecond);
-BENCHMARK(node256_sequential_add)
+BENCHMARK_TEMPLATE(grow_node48_to_node256_sequentially, unodb::olc_db)
+    ->Range(2, 2048)
+    ->Unit(benchmark::kMicrosecond);
+
+BENCHMARK_TEMPLATE(grow_node48_to_node256_randomly, unodb::db)
+    ->Range(2, 2048)
+    ->Unit(benchmark::kMicrosecond);
+BENCHMARK_TEMPLATE(grow_node48_to_node256_randomly, unodb::mutex_db)
+    ->Range(2, 2048)
+    ->Unit(benchmark::kMicrosecond);
+BENCHMARK_TEMPLATE(grow_node48_to_node256_randomly, unodb::olc_db)
+    ->Range(2, 2048)
+    ->Unit(benchmark::kMicrosecond);
+
+BENCHMARK_TEMPLATE(node256_sequential_add, unodb::db)
     ->Range(1, 1024)
     ->Unit(benchmark::kMicrosecond);
-BENCHMARK(node256_random_add)->Range(1, 1024)->Unit(benchmark::kMicrosecond);
-BENCHMARK(minimal_node256_tree_full_scan)
+BENCHMARK_TEMPLATE(node256_sequential_add, unodb::mutex_db)
+    ->Range(1, 1024)
+    ->Unit(benchmark::kMicrosecond);
+BENCHMARK_TEMPLATE(node256_sequential_add, unodb::olc_db)
+    ->Range(1, 1024)
+    ->Unit(benchmark::kMicrosecond);
+
+BENCHMARK_TEMPLATE(node256_random_add, unodb::db)
+    ->Range(1, 1024)
+    ->Unit(benchmark::kMicrosecond);
+BENCHMARK_TEMPLATE(node256_random_add, unodb::mutex_db)
+    ->Range(1, 1024)
+    ->Unit(benchmark::kMicrosecond);
+BENCHMARK_TEMPLATE(node256_random_add, unodb::olc_db)
+    ->Range(1, 1024)
+    ->Unit(benchmark::kMicrosecond);
+
+BENCHMARK_TEMPLATE(minimal_node256_tree_full_scan, unodb::db)
     ->Range(4, 4096)
     ->Unit(benchmark::kMicrosecond);
-BENCHMARK(minimal_node256_tree_random_gets)
+BENCHMARK_TEMPLATE(minimal_node256_tree_full_scan, unodb::mutex_db)
     ->Range(4, 4096)
     ->Unit(benchmark::kMicrosecond);
-BENCHMARK(full_node256_tree_full_scan)
+BENCHMARK_TEMPLATE(minimal_node256_tree_full_scan, unodb::olc_db)
+    ->Range(4, 4096)
+    ->Unit(benchmark::kMicrosecond);
+
+BENCHMARK_TEMPLATE(minimal_node256_tree_random_gets, unodb::db)
+    ->Range(4, 4096)
+    ->Unit(benchmark::kMicrosecond);
+BENCHMARK_TEMPLATE(minimal_node256_tree_random_gets, unodb::mutex_db)
+    ->Range(4, 4096)
+    ->Unit(benchmark::kMicrosecond);
+BENCHMARK_TEMPLATE(minimal_node256_tree_random_gets, unodb::olc_db)
+    ->Range(4, 4096)
+    ->Unit(benchmark::kMicrosecond);
+
+BENCHMARK_TEMPLATE(full_node256_tree_full_scan, unodb::db)
     ->Range(128, 131064)
     ->Unit(benchmark::kMicrosecond);
-BENCHMARK(full_node256_tree_random_gets)
+BENCHMARK_TEMPLATE(full_node256_tree_full_scan, unodb::mutex_db)
     ->Range(128, 131064)
     ->Unit(benchmark::kMicrosecond);
-BENCHMARK(full_node256_tree_sequential_delete)
+BENCHMARK_TEMPLATE(full_node256_tree_full_scan, unodb::olc_db)
+    ->Range(128, 131064)
+    ->Unit(benchmark::kMicrosecond);
+
+BENCHMARK_TEMPLATE(full_node256_tree_random_gets, unodb::db)
+    ->Range(128, 131064)
+    ->Unit(benchmark::kMicrosecond);
+BENCHMARK_TEMPLATE(full_node256_tree_random_gets, unodb::mutex_db)
+    ->Range(128, 131064)
+    ->Unit(benchmark::kMicrosecond);
+BENCHMARK_TEMPLATE(full_node256_tree_random_gets, unodb::olc_db)
+    ->Range(128, 131064)
+    ->Unit(benchmark::kMicrosecond);
+
+BENCHMARK_TEMPLATE(full_node256_tree_sequential_delete, unodb::db)
     ->Range(192, 196608)
     ->Unit(benchmark::kMicrosecond);
-BENCHMARK(full_node256_tree_random_delete)
+BENCHMARK_TEMPLATE(full_node256_tree_sequential_delete, unodb::mutex_db)
+    ->Range(192, 196608)
+    ->Unit(benchmark::kMicrosecond);
+BENCHMARK_TEMPLATE(full_node256_tree_sequential_delete, unodb::olc_db)
+    ->Range(192, 196608)
+    ->Unit(benchmark::kMicrosecond);
+
+BENCHMARK_TEMPLATE(full_node256_tree_random_delete, unodb::db)
+    ->Range(192, 196608)
+    ->Unit(benchmark::kMicrosecond);
+BENCHMARK_TEMPLATE(full_node256_tree_random_delete, unodb::mutex_db)
+    ->Range(192, 196608)
+    ->Unit(benchmark::kMicrosecond);
+BENCHMARK_TEMPLATE(full_node256_tree_random_delete, unodb::olc_db)
     ->Range(192, 196608)
     ->Unit(benchmark::kMicrosecond);
